Genesis ledger/block reconstruction + proving

+++ ledger/src/scan_state/pending_coinbase.rs
@@ -922,12 +922,12 @@ impl PendingCoinbaseWitness {
     }
 
     fn set_oldest_coinbase_stack(&mut self, idx: Address, stack: Stack) {
-        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth as usize;
+        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth;
         self.pending_coinbase.set_stack(depth, idx, stack, false);
     }
 
     fn find_index_of_newest_stacks(&self) -> (Address, Address) {
-        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth as usize;
+        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth;
 
         let index1 = {
             let stack_id = self.pending_coinbase.latest_stack_id(self.is_new_stack);
@@ -958,7 +958,7 @@ impl PendingCoinbaseWitness {
     }
 
     fn set_coinbase_stack(&mut self, idx: Address, stack: Stack) {
-        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth as usize;
+        let depth = CONSTRAINT_CONSTANTS.pending_coinbase_depth;
         self.pending_coinbase
             .set_stack(depth, idx, stack, self.is_new_stack);
     }
